+++ src/HeaderV2/ProductSwitch/ProductSwitchV2.styles.ts
@@ -16,11 +16,3 @@ export const ItemWrapper = styled.div`
   display: flex;
   flex-direction: column;
 `;
-
-export const ItemTitle = styled.span`
-  font-family: ${({ theme }) => theme.typography.fontFamily.sans};
-  font-size: ${({ theme }) => theme.typography.scale.sm.size};
-  line-height: ${({ theme }) => theme.typography.scale.sm.lineHeight};
-  font-weight: ${({ theme }) => theme.typography.fontWeight.medium};
-  color: ${({ theme }) => theme.colors.indigo[800]};
-`;